refactor: modularize CLIcore and fps_processcmdline
